requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.3
cssselect==1.2.0
dateparser==1.2.0
dill==0.3.8
dnspython==2.5.0
email-validator==2.1.0.post1
fake-useragent==1.4.0
fastapi==0.109.2
fastapi-cache2==0.2.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httptools==0.6.1
httpx==0.13.3
huggingface-hub==0.20.3
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
kombu==5.3.5
lxml==5.1.0
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
multiprocess==0.70.16
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
orjson==3.9.14
packaging==23.2
pathos==0.3.2
pendulum==3.0.0
pillow==10.2.0
pluggy==1.4.0
pox==0.3.4
ppft==1.7.6.8
prompt-toolkit==3.0.43
pycparser==2.21
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyExecJS==1.5.1
pytest==8.0.1
pytest-asyncio==0.23.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
rfc3986==1.5.0
safetensors==0.4.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
starlette==0.36.3
time-machine==2.13.0
tinysegmenter==0.3
tldextract==5.1.1
tokenizers==0.15.2
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.0
uvicorn==0.27.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
gunicorn==21.2.0